Legger inn kodeverk i innsok historikk
