+++ apps/innbyggers-flate/src/pages/DeltakerPage.tsx
@@ -37,7 +37,7 @@ export const DeltakerPage = () => {
   const tiltakOgStedTekst = hentTiltakHosArrangorTittel(
     deltaker.deltakerliste.tiltakskode,
     deltaker.deltakerliste.arrangorNavn,
-    deltaker.deltakerliste.kodeverk
+    deltaker.deltakerliste.opplaringKategoriseringValg
   )
   const skalViseDato =
     deltaker.startdato &&
@@ -157,7 +157,7 @@ export const DeltakerPage = () => {
       <DeltakelseInnhold
         tiltakskode={deltaker.deltakerliste.tiltakskode}
         deltakelsesinnhold={deltaker.deltakelsesinnhold}
-        kodeverk={deltaker.deltakerliste.kodeverk}
+        kodeverk={deltaker.deltakerliste.opplaringKategoriseringValg}
         heading={
           <Heading level="2" size="medium" className="mt-8">
             Dette er innholdet
