+++ apps/nav-veileders-flate/src/api/data/deltaker.ts
@@ -46,7 +46,7 @@ export const deltakerlisteSchema = z.object({
   erEnkeltplass: z.boolean(),
   oppmoteSted: z.string().nullable(),
   pameldingstype: z.enum(Pameldingstype),
-  kodeverk: opplaringKategoriseringSchema.nullable(),
+  opplaringKategoriseringValg: opplaringKategoriseringSchema.nullable(),
   prisinformasjon: prisinformasjonSchema.nullish()
 })
