+++ apps/nav-veileders-flate/src/components/pamelding/handlinger/meld-pa-direkte/MeldPaDirekteModalEnkeltPlass.tsx
@@ -41,7 +41,7 @@ export const MeldPaDirekteModalEnkeltPlass = ({ open, onClose }: Props) => {
   const tiltakHosArrangorTekst = hentTiltakHosArrangorTittel(
     deltakerliste.tiltakskode,
     deltakerliste.arrangorNavn,
-    deltakerliste.kodeverk
+    deltakerliste.opplaringKategoriseringValg
   )
   const returnToFrontpageWithSuccessMessage = () => {
     doRedirect(DELTAKELSESOVERSIKT_LINK, {
