+++ apps/nav-veileders-flate/src/components/utkast/UtkastDeltakerEnkeltplass.tsx
@@ -23,7 +23,7 @@ export const UtkastDeltakerEnkeltplass = () => {
           tiltakskode,
           deltaker.deltakerliste.deltakerlisteNavn,
           deltaker.deltakerliste.arrangorNavn,
-          deltaker.deltakerliste.kodeverk
+          deltaker.deltakerliste.opplaringKategoriseringValg
         )}
       />
       <BodyLong size="small" className="mt-2 whitespace-pre-wrap">
@@ -34,7 +34,7 @@ export const UtkastDeltakerEnkeltplass = () => {
       <DeltakelseInnhold
         tiltakskode={tiltakskode}
         deltakelsesinnhold={deltaker.deltakelsesinnhold}
-        kodeverk={deltaker.deltakerliste.kodeverk}
+        kodeverk={deltaker.deltakerliste.opplaringKategoriseringValg}
         heading={
           <Heading level="3" size="medium">
             Dette er innholdet
