+++ apps/nav-veileders-flate/src/mocks/MockHandler.ts
@@ -115,10 +115,10 @@ export class MockHandler {
         pameldingstype: Pameldingstype.TRENGER_GODKJENNING,
         oppmoteSted:
           'Fjordgata 7b, 00 Stedet. Inngangsdør rundt svingen. Oppmøte kl. 09:00. ',
-        kodeverk: createMockFlatKodeverk(
+        opplaringKategoriseringValg: createMockFlatKodeverk(
           this.tiltakskode,
           erEnkeltplass
-        ) as DeltakerResponse['deltakerliste']['kodeverk'],
+        ),
         prisinformasjon: null
       },
       status: {
@@ -461,12 +461,10 @@ export class MockHandler {
 
     if (oppdatertPamelding) {
       oppdatertPamelding.deltakerliste.tiltakskode = tiltakskode
-      oppdatertPamelding.deltakerliste.kodeverk = erNyEnkeltplass
-        ? (createMockFlatKodeverk(
-            this.tiltakskode,
-            erNyEnkeltplass
-          ) as DeltakerResponse['deltakerliste']['kodeverk'])
-        : null
+      oppdatertPamelding.deltakerliste.opplaringKategoriseringValg =
+        erNyEnkeltplass
+          ? createMockFlatKodeverk(this.tiltakskode, erNyEnkeltplass)
+          : null
       oppdatertPamelding.adresseDelesMedArrangor =
         delesAdresseMedArrangor(tiltakskode)
 
@@ -824,8 +822,11 @@ export class MockHandler {
   }
 
   getHistorikk() {
-    return HttpResponse.json(
+    const erEnkeltplass = this.pamelding?.deltakerliste.erEnkeltplass
+    const erFellesOppstart =
       this.pamelding?.deltakerliste.oppstartstype === Oppstartstype.FELLES
+    return HttpResponse.json(
+      erFellesOppstart || erEnkeltplass
         ? lagHistorikkFellesOppstart()
         : createHistorikk()
     )
