+++ packages/deltaker-flate-common/index.ts
@@ -31,12 +31,14 @@ export * from './model/endre-deltaker'
 export * from './model/forslag'
 export * from './model/utils'
 export * from './model/kodeverk'
+export * from './model/kodeverkResponse'
 export * from './model/prisinformasjon'
 export * from './utils/displayText'
 export * from './utils/constants'
 export * from './utils/utils'
 export * from './utils/kodeverkUtils'
 export * from './utils/mockData'
+export * from './utils/mockKodeverk'
 export * from './utils/forslagUtils'
 export * from './utils/faro'
 export * from './components/ErrorFallback'
